pyproject.toml
[tool.poetry]
name = "kvmp"
version = "0.1.0"
description = "KVM manager"
authors = ["Recursive Labs LTD"]
license = "Copyright (C) 2022"
readme = "README.md"
# packages = [{include = "kvmp"}]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
records = "0.5.3"
psycopg2 = "^2.9.5"
libvirt-python = "8.9.0"
pytailwindcss = "^0.1.4"
flask-seasurf = "^1.1.1"
flask-login = "^0.6.2"
werkzeug = "^2.2.2"
paramiko = "^2.12.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"
black = "^21.12b0"
pytest-watch = "^4.2.0"
# commitizen = "^2.20.4"
# pre-commit = "^2.17.0"

[tool.poetry.group.dev.dependencies]
pytailwindcss = "^0.1.4"
flask-assets = "^2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  ".",
  "./kvmp"
]
